+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **TimeSeries**, a sequence of data points indexed by time.
 - **BooleanTimeSeries**, a time-series of boolean data.
 - **NumericTimeSeries**, a time-series of numeric data.
+- **TimeUnit**, an enumerator of time-units.